Extract server cert generator
