+++ internal/webhookcert/clock.go
@@ -14,3 +14,11 @@ type realClock struct {
 func (realClock) now() time.Time {
 	return time.Now().UTC()
 }
+
+type mockClock struct {
+	t time.Time
+}
+
+func (c mockClock) now() time.Time {
+	return c.t
+}
